chore(db,model): Move NvlinkNmxcEndpoint to model
