+++ crates/api-model/src/rpc_conv/mod.rs
@@ -45,6 +45,7 @@ pub mod network_devices;
 pub mod network_prefix;
 pub mod network_security_group;
 pub mod network_segment;
+pub mod nmxc;
 pub mod nvl_logical_partition;
 pub mod nvl_partition;
 pub mod operating_system_definition;
